pyproject.toml
[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = "A base Flask project template"
authors = [""]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
flask = "3.0.2"
gunicorn = "21.2.0"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
numpy = "^1.26.4"
requests = "^2.31.0"
flask-bootstrap = "^3.3.7.1"
scikit-surprise = "^1.1.3"
ipywidgets = "^8.1.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["W291", "W292", "W293"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"